Add methods arity and foldLeft0 (#261)
* Added methods arity and foldLeft0

I am not sure foldLeft0 is the perfect or desired interface. But I am
opening this PR to show that (at least to my understanding) a TypeClass
like ShowType can not be implemented in the current interface.

Of course one can argue that the traversal in ShowType should be done at
compile time since it does not depend on any runtime types. But I have a
use case where I want to have a type class that need to traverse the
elements in multiple ways and being able to implement all of them using
the shapeless 3 API would be useful.

* Add erased sealed traits to excluded MiMa filters

* Add the rest of foltLeft0 syntax

---------

Co-authored-by: Georgi Krastev <joro.kr.21@gmail.com>