Fix some of the scaladoc links
If ones removes the line `Compile / doc / sources := Nil,` from the
build and run `sbt doc` one get warnings like:
```
[warn] -- Warning: modules/deriving/src/main/scala/shapeless3/deriving/kinds.scala:66:7
[warn] 66 |  type Kind[M <: Mirror, O <: Up] = (M of O) { type Kind = self.type }
[warn]    |       ^
[warn]    |       Couldn't resolve a member for the given link query: Mirror
```
Based on the dicussion in https://github.com/scala/scala3/issues/20028
my understanding is that this means that scaldoc can not resolve the
existing link.

The reason I want to fix this is that it will also show up for users
implementing custom `Kinds`. E.g in some private project where I extends
Kind I get warnings like this in my scaladoc:
```
[warn] -- Warning: modules/deriving/src/main/scala/shapeless3/deriving/kinds.scala:66:7
[warn] Couldn't resolve a member for the given link query: Mirror
```
(the error changed slighlty in the latest Scala 3 version). Since I
don't think this is solvable for a user I suggest that we just make the
links absolute which seems to solve the warnings.